configure vcr to record new episodes instead of once
I had to change the default configuration that VCR uses to record due to
a possible conflict between VCR and Capybara. I have searched a little
and it seems that others had the same problem before.

Or it could be a simple intermittent problem with contentful_model that
I didn't have before.